feat(microservices): Removes redundant kafka external interface
